WIP: Add basic enemy movement
